refactor: align project module with feature-based architecture
